+++ pipetune/activation/__init__.py
@@ -8,6 +8,7 @@ from pipetune.activation.installer import (
 )
 from pipetune.activation.rollback import render_rollback_result, rollback_profile
 from pipetune.activation.status import render_activation_status, render_installed_profiles
+from pipetune.activation.state import build_state_doctor_report, render_state_doctor_report
 
 __all__ = [
     "install_profile",
@@ -16,6 +17,8 @@ __all__ = [
     "render_install_result",
     "render_installed_profiles",
     "render_rollback_result",
+    "build_state_doctor_report",
+    "render_state_doctor_report",
     "rollback_profile",
     "run_install_dry_run",
 ]